requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.6.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
distlib==0.3.5
Django==4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.7.1
flake8==5.0.3
identify==2.5.2
interrogate==1.5.0
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.0
pyflakes==2.5.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tabulate==0.8.10
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
virtualenv==20.16.2
